# services/core-schedule/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "core-schedule",
    "description": "Schedule service — university timetable, templates, overrides",
    "version": "0.1.0"
  },
  "paths": {
...
    "/groups/{group_name}/schedule": {
      "get": {
        "tags": ["groups"],
        "summary": "Group Schedule",
        "operationId": "group_schedule_groups__group_name__schedule_get",
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Name"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date To"
            }
          },
          {
            "name": "semester_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DaySchedule"
                  },
                  "title": "Response Group Schedule Groups  Group Name  Schedule Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_name}/exams": {
      "get": {
        "tags": ["groups"],
        "summary": "Group Exams",
        "operationId": "group_exams_groups__group_name__exams_get",
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Name"
            }
          },
          {
            "name": "semester_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionEventResponse"
                  },
                  "title": "Response Group Exams Groups  Group Name  Exams Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_name}/template": {
      "get": {
        "tags": ["groups"],
        "summary": "Group Template",
        "operationId": "group_template_groups__group_name__template_get",
        "parameters": [
          {
            "name": "group_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Name"
            }
          },
          {
            "name": "semester_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/teachers/{teacher_name}/template": {
      "get": {
        "tags": ["teachers"],
        "summary": "Teacher Template",
        "operationId": "teacher_template_teachers__teacher_name__template_get",
        "parameters": [
          {
            "name": "teacher_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Teacher Name"
            }
          },
          {
            "name": "semester_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/templates": {
      "get": {
        "tags": ["admin-templates"],
        "summary": "List Templates",
        "operationId": "list_templates_admin_templates_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "semester_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Semester Id"
            }
          },
          {
            "name": "week_number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Week Number"
            }
          },
          {
            "name": "day_of_week",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day Of Week"
            }
          },
          {
            "name": "group_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group Name"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateSlotResponse"
                  },
                  "title": "Response List Templates Admin Templates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["admin-templates"],
        "summary": "Create Template",
        "operationId": "create_template_admin_templates_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateSlotCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateSlotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/templates/bulk": {
      "post": {
        "tags": ["admin-templates"],
        "summary": "Create Templates Bulk",
        "operationId": "create_templates_bulk_admin_templates_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateSlotBulkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TemplateSlotResponse"
                  },
                  "type": "array",
                  "title": "Response Create Templates Bulk Admin Templates Bulk Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/templates/{template_id}": {
      "put": {
        "tags": ["admin-templates"],
        "summary": "Update Template",
        "operationId": "update_template_admin_templates__template_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateSlotUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateSlotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["admin-templates"],
        "summary": "Delete Template",
        "operationId": "delete_template_admin_templates__template_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Template Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DaySchedule": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "day_name": {
            "type": "string",
            "title": "Day Name"
          },
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonSlot"
            },
            "type": "array",
            "title": "Lessons"
          }
        },
        "type": "object",
        "required": ["date", "day_of_week", "day_name", "lessons"],
        "title": "DaySchedule"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LessonSlot": {
        "properties": {
          "slot_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Number"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "lesson_type": {
            "type": "string",
            "title": "Lesson Type"
          },
          "teacher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Name"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "classroom_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classroom Name"
          },
          "is_override": {
            "type": "boolean",
            "title": "Is Override",
            "default": false
          },
          "override_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Comment"
          }
        },
        "type": "object",
        "required": [
          "slot_number",
          "start_time",
          "end_time",
          "subject_name",
          "lesson_type",
          "teacher_name",
          "group_name",
          "classroom_name"
        ],
        "title": "LessonSlot"
      },
...
      "SessionEventResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "semester_id": {
            "type": "string",
            "format": "uuid",
            "title": "Semester Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "slot_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Number"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "lesson_type": {
            "type": "string",
            "title": "Lesson Type"
          },
          "teacher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Name"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "classroom_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classroom Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "semester_id",
          "date",
          "slot_number",
          "start_time",
          "end_time",
          "subject_name",
          "lesson_type",
          "teacher_name",
          "group_name",
          "classroom_name"
        ],
        "title": "SessionEventResponse"
      },
...
      "TemplateResponse": {
        "properties": {
          "semester_id": {
            "type": "string",
            "format": "uuid",
            "title": "Semester Id"
          },
          "semester_name": {
            "type": "string",
            "title": "Semester Name"
          },
          "week_1": {
            "items": {
              "$ref": "#/components/schemas/DayTemplate"
            },
            "type": "array",
            "title": "Week 1"
          },
          "week_2": {
            "items": {
              "$ref": "#/components/schemas/DayTemplate"
            },
            "type": "array",
            "title": "Week 2"
          }
        },
        "type": "object",
        "required": ["semester_id", "semester_name", "week_1", "week_2"],
        "title": "TemplateResponse"
      },
      "TemplateSlotBulkCreate": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TemplateSlotCreate"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": ["items"],
        "title": "TemplateSlotBulkCreate"
      },
      "TemplateSlotCreate": {
        "properties": {
          "semester_id": {
            "type": "string",
            "format": "uuid",
            "title": "Semester Id"
          },
          "week_number": {
            "type": "integer",
            "title": "Week Number"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "slot_number": {
            "type": "integer",
            "title": "Slot Number"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "lesson_type": {
            "type": "string",
            "title": "Lesson Type"
          },
          "teacher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Name"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "classroom_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classroom Name"
          }
        },
        "type": "object",
        "required": [
          "semester_id",
          "week_number",
          "day_of_week",
          "slot_number",
          "start_time",
          "end_time",
          "subject_name",
          "lesson_type",
          "group_name"
        ],
        "title": "TemplateSlotCreate"
      },
      "TemplateSlotResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "semester_id": {
            "type": "string",
            "format": "uuid",
            "title": "Semester Id"
          },
          "week_number": {
            "type": "integer",
            "title": "Week Number"
          },
          "day_of_week": {
            "type": "integer",
            "title": "Day Of Week"
          },
          "slot_number": {
            "type": "integer",
            "title": "Slot Number"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "subject_name": {
            "type": "string",
            "title": "Subject Name"
          },
          "lesson_type": {
            "type": "string",
            "title": "Lesson Type"
          },
          "teacher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Name"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "classroom_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classroom Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "semester_id",
          "week_number",
          "day_of_week",
          "slot_number",
          "start_time",
          "end_time",
          "subject_name",
          "lesson_type",
          "teacher_name",
          "group_name",
          "classroom_name"
        ],
        "title": "TemplateSlotResponse"
      },
      "TemplateSlotUpdate": {
        "properties": {
          "week_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week Number"
          },
          "day_of_week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day Of Week"
          },
          "slot_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Number"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "subject_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Name"
          },
          "lesson_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesson Type"
          },
          "teacher_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Teacher Name"
          },
          "group_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Name"
          },
          "classroom_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classroom Name"
          }
        },
        "type": "object",
        "title": "TemplateSlotUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}